+++ test/line_style.cxx
@@ -47,10 +47,10 @@ void test_box::draw() {
   dashes[3] = char(sliders[8]->value());
   dashes[4] = 0;
   fl_line_style(
-    choice[0]->mvalue()->argument() +
+    (int)(choice[0]->mvalue()->argument() +
     choice[1]->mvalue()->argument() +
-    choice[2]->mvalue()->argument(),
-    long(sliders[3]->value()),          // width
+    choice[2]->mvalue()->argument()),
+    (int)(sliders[3]->value()),          // width
     dashes);
 
   // draw the defined fl_rect and fl_vertex first and then
